Fix issue where MySQL does not start on system reboot
The reason being that the "remote-fs.target" service is depending on
GlusterFS to start, but that is started by Monit, which takes a while
until it runs the first 5-second cycle.

Now, the "mysql.service" service still depends on "remote-fs.target"
(Note the hard Requires= in the default MySQL systemd file), which means
systemd will skip it, as its depedencies are NOT satisfied.

The fix: just avoid "remote-fs.target" dependency, as the DB is on "/".